.pub-cache/hosted/pub.dartlang.org/html-0.15.0/test/data/tree-construction/tests4.dat

#data
direct div content
#errors
#document-fragment
div
#document
| "direct div content"

#data
direct textarea content
#errors
#document-fragment
textarea
#document
| "direct textarea content"

#data
textarea content with <em>pseudo</em> <foo>markup
#errors
#document-fragment
textarea
#document
| "textarea content with <em>pseudo</em> <foo>markup"

#data
this is &#x0043;DATA inside a <style> element
#errors
#document-fragment
style
#document
| "this is &#x0043;DATA inside a <style> element"

#data
</plaintext>
#errors
#document-fragment
plaintext
#document
| "</plaintext>"

#data
setting html's innerHTML
#errors
Line: 1 Col: 24 Unexpected EOF in inner html mode.
#document-fragment
html
#document
| <head>
| <body>
|   "setting html's innerHTML"

#data
<title>setting head's innerHTML</title>
#errors
#document-fragment
head
#document
| <title>
|   "setting head's innerHTML"
